PythonMLModel/requirements.txt
gradio==3.39.0
Pillow==9.4.0
facenet-pytorch==2.5.2
torch==1.11.0
opencv-python==4.7.0.72
grad-cam==1.4.6
